split out keymap_legacy
